+++ STS2AIAgent.Tests/TestRunner.cs
@@ -311,6 +311,12 @@ internal static class TestRunner
         yield return ("CombatDiagnostics.CancelPlayCard", () => Task.Run(CombatDiagnosticsContractTests.PlayCardTimeoutCancelsNativeGameAction));
         yield return ("CombatDiagnostics.OwnPets", () => Task.Run(CombatDiagnosticsContractTests.CombatPayloadExposesOwnPets));
         yield return ("ProfileSelection.NativeSwitch", () => Task.Run(ProfileSelectionContractTests.NativeProfileIdentityAndSwitchAreWiredEndToEnd));
+        yield return ("DecisionLog.BoundsAndRedacts", () => Task.Run(DecisionLogTests.Record_RedactsBoundsAndKeepsNewest));
+        yield return ("DecisionLog.PersistsAndRotates", () => Task.Run(DecisionLogTests.Record_PersistsJsonlAndRotates));
+        yield return ("DecisionLog.PersistenceFailureIsSafe", () => Task.Run(DecisionLogTests.Record_UnwritablePathNeverThrows));
+        yield return ("DecisionLog.ClientContextReason", () => Task.Run(DecisionContextTests.ClientContext_ReasonIsOptionalAndTrimmed));
+        yield return ("Mcp.DecisionLogTool", McpServiceTests.ToolsCall_DecisionLogRecordsAcceptedActOnly);
+        yield return ("Mcp.DecisionLogAbsentIsEmpty", McpServiceTests.NativeServerWithoutDecisionLog_StaysSilent);
         yield return ("AgentLoop.PlayOnce", AgentLoopTests.PlayOnce_ExecutesSingleValidatedAct);
         yield return ("AgentLoop.CrystalArgs", AgentLoopTests.PlayOnce_ForwardsCrystalSphereArguments);
         yield return ("AgentTools.CrystalSchema", () => Task.Run(AgentLoopTests.ActToolSchema_IncludesCrystalSphereArguments));